+++ lib/espn_box_score/espn_new.rb
@@ -1,0 +1,4 @@
+module EspnBoxScore
+  class EspnNew < Base
+  end
+end
