+++ lib/espn_box_score/nba.rb
@@ -1,0 +1,4 @@
+module EspnBoxScore
+  class Nba < EspnOld
+  end
+end
