+++ lib/espn_box_score/ncb.rb
@@ -1,0 +1,4 @@
+module EspnBoxScore
+  class Ncb < EspnOld
+  end
+end
