+++ lib/espn_box_score/ncw.rb
@@ -1,0 +1,4 @@
+module EspnBoxScore
+  class Ncw < EspnOld
+  end
+end
